+++ src/Node.java
@@ -3,7 +3,7 @@ import java.util.ArrayList;
 
 public class Node implements Comparable<Node>{
 	private String name;
-	private int minDistace = 2000000; 
+	private double minDistace = 2000000; 
 	//This is the minimum distance it takes to move from the origin node to the current node.
 	//Use 2,147,483,647 to represent infinity. (We cannot reach this node directly.)
 	//Biggest int in java is that anyway.
@@ -21,11 +21,11 @@ public class Node implements Comparable<Node>{
 		return prev;
 	}
 	
-	public void setMinDist(int dist){
+	public void setMinDist(double dist){
 		minDistace = dist;
 	}
 	
-	public int getMinDist(){
+	public double getMinDist(){
 		return minDistace;
 	}
 	
@@ -34,6 +34,6 @@ public class Node implements Comparable<Node>{
 	}
 
 	public int compareTo(Node other) {
-		return Integer.compare(minDistace, other.minDistace);
+		return Double.compare(minDistace, other.minDistace);
 	}
 }
